Fixed 'Save to EmailLater' Folder
The issue had to do with the fact that I did not create an object of
EmailPrompt but instead used The parent to check the status ofthe 'email
later' checkbox. It never changed and since it was always unchecked when
I used it, I never noticed that it was checking the wrong object.